Put interfaces TextFormat & Converter to separate files
