+++ bin/builder/color_text_format/converters/json_converter.dart
@@ -1,5 +1,5 @@
 import '../formats/json_format.dart';
-import '../color_reader/color_text_reader.dart';
+import 'converter.dart';
 
 class JsonConverter extends Converter<JsonFormat> {
   @override
