+++ bin/builder/color_text_format/formats/text_format.dart
@@ -1,0 +1,9 @@
+/// Product
+abstract class TextFormat {
+  String get content;
+
+  @override
+  String toString() => '$runtimeType(\n'
+      '$content'
+      '\n)';
+}
