+++ bin/builder/color_text_format/main.dart
@@ -1,7 +1,7 @@
+import 'color_reader/color_text_reader.dart';
 import 'converters/html_converter.dart';
 import 'converters/json_converter.dart';
 import 'converters/console_converter.dart';
-import 'color_reader/color_text_reader.dart';
 
 void main() {
   final reader = ColorTextReader(
